requirements.txt
functions-framework==3.*
google-cloud-aiplatform >= 1.31.0
google-cloud-logging
python-dotenv
